requirements.txt
flask
groq
pymilvus
sentence-transformers
openai-whisper
python-dotenv
werkzeug